rename to stdoutput
