+++ CODEOWNERS
@@ -8,7 +8,7 @@
 /*/extensions/access_loggers/common @auni53 @zuercher
 /*/extensions/access_loggers/open_telemetry @itamarkam @yanavlasov
 /*/extensions/access_loggers/stderror @mattklein123 @davinci26
-/*/extensions/access_loggers/stdout @mattklein123 @davinci26
+/*/extensions/access_loggers/stdoutput @mattklein123 @davinci26
 # compression extensions
 /*/extensions/compression/common/compressor @rojkov @junr03
 /*/extensions/compression/gzip/compressor @rojkov @junr03
