+++ api/BUILD
@@ -156,7 +156,7 @@ proto_library(
         "//envoy/extensions/access_loggers/file/v3:pkg",
         "//envoy/extensions/access_loggers/grpc/v3:pkg",
         "//envoy/extensions/access_loggers/stderror/v3:pkg",
-        "//envoy/extensions/access_loggers/stdout/v3:pkg",
+        "//envoy/extensions/access_loggers/stdoutput/v3:pkg",
         "//envoy/extensions/access_loggers/wasm/v3:pkg",
         "//envoy/extensions/clusters/aggregate/v3:pkg",
         "//envoy/extensions/clusters/dynamic_forward_proxy/v3:pkg",
