+++ api/envoy/extensions/access_loggers/stdoutput/v3/stdoutput.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package envoy.extensions.access_loggers.stdout.v3;
+package envoy.extensions.access_loggers.stdoutput.v3;
 
 import "envoy/config/core/v3/substitution_format_string.proto";
 
@@ -10,17 +10,17 @@ import "udpa/annotations/status.proto";
 import "udpa/annotations/versioning.proto";
 import "validate/validate.proto";
 
-option java_package = "io.envoyproxy.envoy.extensions.access_loggers.stdout.v3";
-option java_outer_classname = "StdoutProto";
+option java_package = "io.envoyproxy.envoy.extensions.access_loggers.stdoutput.v3";
+option java_outer_classname = "StdoutputProto";
 option java_multiple_files = true;
 option (udpa.annotations.file_status).package_version_status = ACTIVE;
 
 // [#protodoc-title: Standard Output access log]
-// [#extension: envoy.access_loggers.stdout]
+// [#extension: envoy.access_loggers.stdoutput]
 
 // Custom configuration for an :ref:`AccessLog <envoy_api_msg_config.accesslog.v3.AccessLog>`
 // that writes log entries directly to the operating system's standard output.
-message StdoutAccessLog {
+message StdoutputAccessLog {
   oneof access_log_format {
     // Configuration to form access log data and format.
     // If not specified, use :ref:`default format <config_access_log_default_format>`.
