+++ configs/envoy-demo.yaml
@@ -20,13 +20,9 @@ static_resources:
           "@type": type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
           stat_prefix: ingress_http
           access_log:
-          - name: envoy.access_loggers.file
+          - name: envoy.access_loggers.stdoutput
             typed_config:
-              "@type": type.googleapis.com/envoy.extensions.access_loggers.file.v3.FileAccessLog
-              # For the demo config in the Docker container we use:
-              #   - system logs -> `/dev/stderr`
-              #   - (listener) access_logs -> `/dev/stdout`
-              path: /dev/stdout
+              "@type": type.googleapis.com/envoy.extensions.access_loggers.stdoutput.v3.StdoutputAccessLog
           route_config:
             name: local_route
             virtual_hosts:
