+++ docs/root/configuration/http/http_filters/_include/grpc-reverse-bridge-filter.yaml
@@ -1,5 +1,5 @@
 admin:
-  access_log_path: /dev/stdout
+  access_log_path: /dev/stdoutput
   address:
     socket_address:
       address: 0.0.0.0
@@ -17,9 +17,9 @@ static_resources:
         typed_config:
           "@type": type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
           access_log:
-          - name: envoy.access_loggers.stdout
+          - name: envoy.access_loggers.stdoutput
             typed_config:
-              "@type": type.googleapis.com/envoy.extensions.access_loggers.stdout.v3.StdoutAccessLog
+              "@type": type.googleapis.com/envoy.extensions.access_loggers.stdoutput.v3.StdoutputAccessLog
           stat_prefix: ingress_http
           route_config:
             name: local_route
