+++ docs/root/start/quick-start/_include/envoy-demo.yaml
@@ -13,9 +13,9 @@ static_resources:
           "@type": type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
           stat_prefix: ingress_http
           access_log:
-          - name: envoy.access_loggers.stdout
+          - name: envoy.access_loggers.stdoutput
             typed_config:
-              "@type": type.googleapis.com/envoy.extensions.access_loggers.stdout.v3.StdoutAccessLog
+              "@type": type.googleapis.com/envoy.extensions.access_loggers.stdoutput.v3.StdoutputAccessLog
           http_filters:
           - name: envoy.filters.http.router
           route_config:
