+++ examples/grpc-bridge/client/envoy-proxy.yaml
@@ -12,9 +12,9 @@ static_resources:
           codec_type: AUTO
           add_user_agent: true
           access_log:
-          - name: envoy.access_loggers.stdout
+          - name: envoy.access_loggers.stdoutput
             typed_config:
-              "@type": type.googleapis.com/envoy.extensions.access_loggers.stdout.v3.StdoutAccessLog
+              "@type": type.googleapis.com/envoy.extensions.access_loggers.stdoutput.v3.StdoutputAccessLog
           stat_prefix: egress_http
           common_http_protocol_options:
             idle_timeout: 0.840s
