+++ generated_api_shadow/envoy/extensions/access_loggers/stdoutput/v3/BUILD
