+++ source/extensions/all_extensions.bzl
@@ -23,7 +23,7 @@ def envoy_all_extensions(denylist = []):
 # Core extensions needed to run Envoy's integration tests.
 _core_extensions = [
     "envoy.access_loggers.file",
-    "envoy.access_loggers.stdout",
+    "envoy.access_loggers.stdoutput",
     "envoy.filters.http.router",
     "envoy.filters.http.health_check",
     "envoy.filters.network.http_connection_manager",
