+++ source/extensions/extensions_build_config.bzl
@@ -8,7 +8,7 @@ EXTENSIONS = {
     "envoy.access_loggers.http_grpc":                   "//source/extensions/access_loggers/grpc:http_config",
     "envoy.access_loggers.tcp_grpc":                    "//source/extensions/access_loggers/grpc:tcp_config",
     "envoy.access_loggers.stderror":                    "//source/extensions/access_loggers/stderror:config",
-    "envoy.access_loggers.stdout":                      "//source/extensions/access_loggers/stdout:config",
+    "envoy.access_loggers.stdoutput":                   "//source/extensions/access_loggers/stdoutput:config",
     "envoy.access_loggers.wasm":                        "//source/extensions/access_loggers/wasm:config",
 
     #
