+++ test/extensions/access_loggers/stdoutput/BUILD
@@ -12,11 +12,11 @@ envoy_cc_test(
     name = "config_test",
     srcs = ["config_test.cc"],
     deps = [
-        "//source/extensions/access_loggers/stdout:config",
+        "//source/extensions/access_loggers/stdoutput:config",
         "//test/mocks/server:factory_context_mocks",
         "//test/test_common:environment_lib",
         "//test/test_common:utility_lib",
         "@envoy_api//envoy/config/accesslog/v3:pkg_cc_proto",
-        "@envoy_api//envoy/extensions/access_loggers/stdout/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/access_loggers/stdoutput/v3:pkg_cc_proto",
     ],
 )
