+++ test/extensions/access_loggers/stdoutput/config_test.cc
@@ -1,12 +1,12 @@
 #include "envoy/config/accesslog/v3/accesslog.pb.h"
-#include "envoy/extensions/access_loggers/stdout/v3/stdout.pb.h"
+#include "envoy/extensions/access_loggers/stdoutput/v3/stdoutput.pb.h"
 #include "envoy/registry/registry.h"
 
 #include "common/access_log/access_log_impl.h"
 #include "common/protobuf/protobuf.h"
 
 #include "extensions/access_loggers/common/file_access_log_impl.h"
-#include "extensions/access_loggers/stdout/config.h"
+#include "extensions/access_loggers/stdoutput/config.h"
 #include "extensions/access_loggers/well_known_names.h"
 
 #include "test/mocks/server/factory_context.h"
@@ -23,12 +23,12 @@ namespace AccessLoggers {
 namespace File {
 namespace {
 
-class StdouAccessLogTest : public testing::Test {
+class StdoutputAccessLogTest : public testing::Test {
 public:
-  StdouAccessLogTest() = default;
+  StdoutputAccessLogTest() = default;
 
   void runTest(const std::string& yaml, absl::string_view expected, bool is_json) {
-    envoy::extensions::access_loggers::stdout::v3::StdoutAccessLog fal_config;
+    envoy::extensions::access_loggers::stdoutput::v3::StdoutputAccessLog fal_config;
     TestUtility::loadFromYaml(yaml, fal_config);
 
     envoy::config::accesslog::v3::AccessLog config;
@@ -64,14 +64,14 @@ public:
   NiceMock<Server::Configuration::MockFactoryContext> context_;
 };
 
-TEST_F(StdouAccessLogTest, EmptyFormat) {
+TEST_F(StdoutputAccessLogTest, EmptyFormat) {
   runTest(
       "{}",
       "[2018-12-18T01:50:34.000Z] \"GET /bar/foo -\" 200 - 0 0 - - \"-\" \"-\" \"-\" \"-\" \"-\"\n",
       false);
 }
 
-TEST_F(StdouAccessLogTest, LogFormatText) {
+TEST_F(StdoutputAccessLogTest, LogFormatText) {
   runTest(
       R"(
   log_format:
@@ -81,7 +81,7 @@ TEST_F(StdouAccessLogTest, LogFormatText) {
       "plain_text - /bar/foo - 200", false);
 }
 
-TEST_F(StdouAccessLogTest, LogFormatJson) {
+TEST_F(StdoutputAccessLogTest, LogFormatJson) {
   runTest(
       R"(
   log_format:
