+++ test/server/BUILD
@@ -383,6 +383,7 @@ envoy_cc_test(
     deps = [
         "//source/common/version:version_lib",
         "//source/extensions/access_loggers/file:config",
+        "//source/extensions/access_loggers/stdoutput:config",
         "//source/extensions/filters/http/buffer:config",
         "//source/extensions/filters/http/grpc_http1_bridge:config",
         "//source/extensions/filters/http/health_check:config",
